requirements.txt
InstructorEmbedding
ctransformers
transformers
langchain
streamlit
pypdfium2
librosa
pyyaml